fix: auto_set_term_color cause errors when vim starts with a dir path
calling redraw! in VimEnter may cause some stored variables to be
cleared. this is a real strange bug? but i can confirm it's redraw!
cause the problem. CursorMoved is triggered after VimEnter, so we can
call redraw! at this time. it can clear some leftover chars, which are
used to set colors in terminals, on the screen too. it doesn't have the
problem which VimEnter has.